an ability to generate a cli from code-migrate's API
