+++ packages/code-migrate/src/index.ts
@@ -1,2 +1,3 @@
 export * from './migrate';
 export * from './runMigration';
+export * from './createCli';
